Add functional tournament service
The tournament service fails to build the mock tournament because the Game
interface allows null Team objects. This needs to be changed to either
DocumentReferences (which are lightweight and can be null in firestore) or
id strings (which can be null in firestore).

DocumentReferences are preferred, but it will mean splitting out firestore
storage interfaces from display objects. That will be a pain.